+++ .flake8
@@ -1,0 +1,8 @@
+[flake8]
+extend-ignore = E203
+max-line-length = 119
+ignore = E203,W503,W504
+exclude =
+    .git
+    __pychache__
+    .tox
