--- instant-sim/const/__init__.py
@@ -1,1 +1,0 @@
-from .const import *
