--- instant-sim/service/types/__init__.py
@@ -1,5 +1,0 @@
-from .simulator import (
-    nGraph
-)
-
-from . import response as resp
