--- instant-sim/service/types/simulator.py
@@ -1,3 +1,0 @@
-import networkx as nx
-
-nGraph = nx.classes.graph.Graph
