+++ instant_sim/const/util.py
@@ -1,0 +1,12 @@
+from typing import Dict
+
+
+def get_neighbor(graph, label) -> Dict[str, float]:
+    res: Dict[str, float] = {}
+    for i in graph[label]:
+        t, v = graph.nodes[i].values()
+        if t not in res.keys():
+            res[t] = v
+        else:
+            res[t] += v
+    return res
