+++ instant_sim/service/types/simulator.py
@@ -1,0 +1,5 @@
+from typing import NewType
+
+import networkx as nx
+
+nGraph = NewType("nGraph", nx.classes.graph.Graph)  # type: ignore
