+++ pyproject.toml
@@ -1,0 +1,78 @@
+[tool.poetry]
+name = "instant-idea-sim"
+version = "0.1.0"
+description = "Idea simulator for instant poc"
+authors = ["takesei"]
+license = "MIT"
+
+[tool.poetry.dependencies]
+python = "^3.9"
+fastapi = "^0.74.1"
+google-python-cloud-debugger = "^2.18"
+google-cloud-profiler = "^3.0.7"
+httptools = "^0.4.0"
+matplotlib = "^3.5.1"
+networkx = "^2.7"
+numpy = "^1.22.2"
+pandas = "^1.4.1"
+pydantic = "^1.9.0"
+redis = "^4.1.4"
+requests = "^2.27.1"
+seaborn = "^0.11.2"
+uvicorn = "^0.17.5"
+uvloop = "^0.16.0"
+
+[tool.poetry.dev-dependencies]
+pytest = "^7.0.1"
+flake8 = "^4.0.1"
+black = "^22.1.0"
+isort = "^5.10.1"
+tox = "^3.24.5"
+mypy = "^0.931"
+types-requests = "^2.27.11"
+types-redis = "^4.1.17"
+
+[tool.black]
+line-length = 119
+exclude = '''
+(
+    .git
+    | __pychache__
+    | .tox
+)
+'''
+
+[tool.mypy]
+ignore_missing_imports = true
+
+[tool.isort]
+profile = "black"
+line_length = 119
+
+[build-system]
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
+
+[tool.tox]
+legacy_tox_ini = """
+[tox]
+skipdist = true
+envlist = py39, flake8-lint
+
+[testenv]
+whitelist_externals = poetry
+skip_install = true
+commands =
+    poetry install
+    poetry run pytest -rsfp
+
+[testenv:lint]
+whitelist_externals = poetry
+skip_install = true
+commands =
+    poetry install
+    poetry run isort .
+    poetry run black .
+    poetry run flake8 .
+    poetry run mypy .
+"""
