--- requirements.txt
@@ -1,14 +1,0 @@
-fastapi
-google-python-cloud-debugger
-google-cloud-profiler
-httptools
-matplotlib
-networkx
-numpy
-pandas
-pydantic
-redis
-requests
-seaborn
-uvicorn
-uvloop
