+++ tox.ini
@@ -1,0 +1,35 @@
+[tox]
+envlist =
+    py39
+    lint
+    strictlint
+
+[testenv]
+deps =
+    -rrequirements.txt
+    pytest
+commands =
+    pytest -rsfp
+
+[testenv:lint]
+deps =
+    black
+    flake8
+    isort
+    mypy
+commands =
+    isort .
+    black .
+    flake8 .
+    mypy .
+
+[testenv:strictlint]
+ignore_errors = true
+deps =
+    bandit
+    flake8
+    mypy
+commands =
+    bandit --exclude ./.tox,./venv,./.venv,./**/tests --recursive .
+    flake8 --ignore=W503 .
+    mypy --strict .
